+++ src/server/hashidregistrar.cpp
@@ -20,7 +20,6 @@
 #include "hashidregistrar.h"
 
 #include "common/concurrent.h"
-#include "common/containerutil.h"
 
 #include "database.h"
 
@@ -142,11 +141,11 @@ namespace PMP::Server
         return result;
     }
 
-    QVector<uint> HashIdRegistrar::getAllIdsLoaded()
+    QList<uint> HashIdRegistrar::getAllIdsLoaded()
     {
         QMutexLocker lock(&_mutex);
 
-        return ContainerUtil::toVector(_ids.keys());
+        return _ids.keys();
     }
 
     QVector<QPair<uint, FileHash>> HashIdRegistrar::getExistingIdsOnly(